ocean_dp/requirements.txt
pytz
geomag
netCDF4
gsw
oceansdb
psycopg2
fuzzywuzzy
python-Levenshtein
python-dateutil
si_format
